pyproject.toml
[tool.poetry]
name = "tab-researcher"
version = "0.1.0"
description = ""
authors = ["Aneesh Anil <aneesh@geektrust.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.313"
openai = "^0.28.1"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
